# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Your Auth API",
    "description": "API de autenticacao e identidade para contas proprietarias, projetos, usuarios finais, sessoes, refresh tokens e API keys de projeto.",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/users/refresh": {
      "post": {
        "tags": [
          "Public Project Users"
        ],
        "summary": "Renova a sessao web de um usuario final",
        "description": "Usa o cookie refresh_token para gerar novos tokens e redefinir cookies HTTP-only.",
        "operationId": "refresh",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "description": "Refresh token HTTP-only recebido no login do usuario final.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessao renovada. Novos cookies access-token e refresh_token sao enviados no header Set-Cookie.",
            "headers": {
              "Set-Cookie": {
                "description": "Define novos cookies HTTP-only access-token e refresh_token.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Cookie refresh_token ausente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalido, expirado ou reutilizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario ou sessao associados ao refresh token nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao validar, armazenar ou gerar tokens.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshTokenCookie": []
          }
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "Account Authentication"
        ],
        "summary": "Renova a sessao web da conta",
        "description": "Usa o cookie refresh_token para gerar novos tokens e redefinir os cookies HTTP-only.",
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "description": "Refresh token HTTP-only recebido no login web.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessao renovada. Novos cookies access-token e refresh_token sao enviados no header Set-Cookie.",
            "headers": {
              "Set-Cookie": {
                "description": "Define novos cookies HTTP-only access-token e refresh_token.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountTokensResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Cookie refresh_token ausente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalido, expirado ou reutilizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Conta associada ao refresh token nao encontrada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao validar, armazenar ou gerar tokens.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshTokenCookie": []
          }
        ]
      }
    },
    "/auth/mobile/refresh": {
      "post": {
        "tags": [
          "Account Authentication"
        ],
        "summary": "Renova a sessao mobile da conta",
        "description": "Usa o refresh token enviado no corpo da requisicao para gerar um novo par de tokens.",
        "operationId": "refreshMobileToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRefreshRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sessao renovada com access token e refresh token retornados no body.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountTokensResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido ou erro de validacao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalido, expirado ou reutilizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Conta associada ao refresh token nao encontrada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao validar, armazenar ou gerar tokens.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/mobile/login": {
      "post": {
        "tags": [
          "Account Authentication"
        ],
        "summary": "Autentica uma conta proprietaria no fluxo mobile",
        "description": "Valida as credenciais da conta e retorna access token e refresh token no corpo da resposta.",
        "operationId": "mobileLogin",
        "parameters": [
          {
            "name": "User-Agent",
            "in": "header",
            "description": "Identificador do client/dispositivo usado para registrar o refresh token.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "YourAuthMobile/1.0"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Conta autenticada com tokens retornados no body.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginMobileResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido ou erro de validacao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais invalidas ou login bloqueado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao gerar tokens da sessao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Account Authentication"
        ],
        "summary": "Autentica uma conta proprietaria no fluxo web",
        "description": "Valida as credenciais da conta, cria uma sessao e define cookies HTTP-only de access token e refresh token.",
        "operationId": "login_1",
        "parameters": [
          {
            "name": "User-Agent",
            "in": "header",
            "description": "Identificador do client/dispositivo usado para registrar o refresh token.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Mozilla/5.0"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Conta autenticada. Os cookies access-token e refresh_token sao enviados no header Set-Cookie.",
            "headers": {
              "Set-Cookie": {
                "description": "Define os cookies HTTP-only access-token e refresh_token.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido ou erro de validacao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais invalidas ou login bloqueado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao gerar tokens da sessao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PhoneDTO": {
        "type": "object",
        "properties": {
          "ddd": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\d{2}"
          },
          "number": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\d{8,9}"
          }
        },
        "required": [
          "ddd",
          "number"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "AccountTokensResponseDTO": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "AccountRefreshRequestDTO": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LoginDTO": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "cpf": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
      "AccountResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDTO"
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneDTO"
          },
          "CPF": {
            "type": "string"
          }
        }
      },
      "AddressDTO": {
        "type": "object",
        "properties": {
          "cep": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\d{8}|\\d{5}-\\d{3}"
          },
          "street": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "neighborhood": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "city": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "state": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2
          },
          "number": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          }
        },
        "required": [
          "cep",
          "city",
          "neighborhood",
          "number",
          "state",
          "street"
        ]
      },
      "LoginMobileResponseDTO": {
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AccountResponseDTO"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT de conta proprietaria enviado no header Authorization: Bearer <token>.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "projectApiKey": {
        "type": "apiKey",
        "description": "API key de projeto usada por clientes terceiros. Tambem pode ser enviada como Authorization: Bearer ya_sk_...",
        "name": "X-API-Key",
        "in": "header"
      },
      "accessTokenCookie": {
        "type": "apiKey",
        "description": "Cookie HTTP-only usado pelo fluxo web para autenticacao.",
        "name": "access-token",
        "in": "cookie"
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "description": "Cookie HTTP-only usado pelo fluxo web para renovar a sessao.",
        "name": "refresh_token",
        "in": "cookie"
      }
    }
  }
}